feat: move TypedEventEmitter to commons package
